requirements.txt
opencv-python==4.5.1.48
streamlit==0.80.0
opencv_contrib_python==4.5.1.48
pandas==1.2.4
numpy==1.20.1
Pillow==8.2.0
